move simple_dense_vector_storage
