+++ lib/segment/src/vector_storage/dense/memmap_dense_vector_storage.rs
@@ -224,9 +224,9 @@ mod tests {
     use crate::fixtures::payload_context_fixture::FixtureIdTracker;
     use crate::id_tracker::IdTracker;
     use crate::types::{PointIdType, QuantizationConfig, ScalarQuantizationConfig};
+    use crate::vector_storage::dense::simple_dense_vector_storage::open_simple_vector_storage;
     use crate::vector_storage::new_raw_scorer;
     use crate::vector_storage::quantized::quantized_vectors::QuantizedVectors;
-    use crate::vector_storage::simple_dense_vector_storage::open_simple_vector_storage;
 
     #[test]
     fn test_basic_persistence() {
