+++ lib/segment/src/vector_storage/dense/mod.rs
@@ -2,3 +2,4 @@ pub mod appendable_mmap_dense_vector_storage;
 mod dynamic_mmap_flags;
 pub mod memmap_dense_vector_storage;
 pub mod mmap_dense_vectors;
+pub mod simple_dense_vector_storage;
